+++ search/string_filter.go
@@ -51,7 +51,7 @@ func StringFilterGet(
 			"total",
 			// Cardinality aggregation returns the count of all buckets. 40000 is the maximum precision threshold,
 			// so we use it to get the most accurate approximation.
-			elastic.NewCardinalityAggregation().Field("claims.string.string").PrecisionThreshold(40000), //nolint:gomnd
+			elastic.NewCardinalityAggregation().Field("claims.string.string").PrecisionThreshold(40000), //nolint:mnd
 		),
 	)
 	searchService = searchService.Size(0).Query(query).Aggregation("string", aggregation)